chore(rwa): move facade types directly in to file
