+++ packages/ui/src/helpers/authenticator/index.ts
@@ -5,9 +5,3 @@ export * from './constants';
 export * from './form';
 export * from './utils';
 export * from './formFields';
-export {
-  AuthenticatorRoute,
-  AuthenticatorSendEventAliases,
-  AuthenticatorServiceContextFacade,
-  AuthenticatorServiceFacade,
-} from './types';
